+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### Added
 - Enable associative matching for string concatenation (#3741)
 
+### Fixed
+- Java: separate import static from regular imports during matching (#3772)
+
 ## [0.63.0](https://github.com/returntocorp/semgrep/releases/tag/v0.63.0) - 08-25-2021
 
 ### Added